Move theory engine model builder to its own class.
