+++ src/theory/theory_model.h
@@ -178,88 +178,6 @@ public:
   //---------------------------- end function values
 };/* class TheoryModel */
 
-/** TheoryEngineModelBuilder class
-  *    This model builder will consult all theories in a theory engine for
-  *    collectModelInfo( ... ) when building a model.
-  */
-class TheoryEngineModelBuilder : public ModelBuilder
-{
-protected:
-  /** pointer to theory engine */
-  TheoryEngine* d_te;
-  typedef std::unordered_map<Node, Node, NodeHashFunction> NodeMap;
-  NodeMap d_normalizedCache;
-  typedef std::unordered_set<Node, NodeHashFunction> NodeSet;
-  std::map< Node, Node > d_constantReps;
-
-  /** process build model */
-  virtual bool preProcessBuildModel(TheoryModel* m);
-  virtual bool processBuildModel(TheoryModel* m);
-  virtual void debugModel( TheoryModel* m ) {}
-  /** normalize representative */
-  Node normalize(TheoryModel* m, TNode r, bool evalOnly);
-  bool isAssignable(TNode n);
-  void checkTerms(TNode n, TheoryModel* tm, NodeSet& cache);
-  void assignConstantRep( TheoryModel* tm, Node eqc, Node const_rep );
-  void addToTypeList( TypeNode tn, std::vector< TypeNode >& type_list, std::map< TypeNode, bool >& visiting );
-  /** is v an excluded codatatype value */
-  bool isExcludedCdtValue( Node v, std::set<Node>* repSet, std::map< Node, Node >& assertedReps, Node eqc );
-  bool isCdtValueMatch( Node v, Node r, Node eqc, Node& eqc_m );
-  /** involves usort */
-  bool involvesUSort( TypeNode tn );
-  bool isExcludedUSortValue( std::map< TypeNode, unsigned >& eqc_usort_count, Node v, std::map< Node, bool >& visited );
-protected:
-  /** assign function f based on the model m. 
-  * This construction is based on "table form". For example:
-  * (f 0 1) = 1
-  * (f 0 2) = 2
-  * (f 1 1) = 3
-  * ...
-  * becomes:
-  * f = (lambda xy. (ite (and (= x 0) (= y 1)) 1 
-  *                 (ite (and (= x 0) (= y 2)) 2
-  *                 (ite (and (= x 1) (= y 1)) 3 ...)))
-  */
-  void assignFunction(TheoryModel* m, Node f);
-  /** assign function f based on the model m. 
-  * This construction is based on "dag form". For example:
-  * (f 0 1) = 1
-  * (f 0 2) = 2
-  * (f 1 1) = 3
-  * ...
-  * becomes:
-  * f = (lambda xy. (ite (= x 0) (ite (= y 1) 1
-  *                              (ite (= y 2) 2 ...))
-  *                 (ite (= x 1) (ite (= y 1) 3 ...)
-  *                              ...))
-  * 
-  * where the above is represented as a directed acyclic graph (dag).
-  * This construction is accomplished by assigning values to (f c) 
-  * terms before f, e.g. 
-  * (f 0) = (lambda y. (ite (= y 1) 1
-  *                    (ite (= y 2) 2 ...))
-  * (f 1) = (lambda y. (ite (= y 1) 3 ...))
-  * where
-  * f = (lambda xy. (ite (= x 0) ((f 0) y)
-  *                 (ite (= x 1) ((f 1) y) ...))
-  */
-  void assignHoFunction(TheoryModel* m, Node f);
-  /** Assign all unassigned functions in the model m (those returned by TheoryModel::getFunctionsToAssign), 
-  * using the two functions above. Currently:
-  * If ufHo is disabled, we call assignFunction for all functions. 
-  * If ufHo is enabled, we call assignHoFunction.
-  */
-  void assignFunctions(TheoryModel* m);
-public:
-  TheoryEngineModelBuilder(TheoryEngine* te);
-  virtual ~TheoryEngineModelBuilder(){}
-  /** Build model function.
-   *    Should be called only on TheoryModels m
-   */
-  bool buildModel(Model* m);
-  void debugCheckModel(Model* m);
-};/* class TheoryEngineModelBuilder */
-
 }/* CVC4::theory namespace */
 }/* CVC4 namespace */
